% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tech-AgriFertilizer.R
\docType{data}
\name{AgriFertilizer}
\alias{AgriFertilizer}
\title{Data Set of Agricultural Fertilizer}
\format{
A data frame:
\describe{
\item{province}{province, including nation}
\item{year}{year, begin with year 2010}
\item{chn_block4}{vars name, in chinese}
\item{value}{value of variables}
\item{units}{units of variables}
\item{variables}{vars name, by coding}
}
}
\source{
\url{https://www.huhuaping.com/}
}
\usage{
AgriFertilizer
}
\description{
A data set containing Agricultural Fertilizer statistics from the Yearbook of Rural,
with long data format.
}
\keyword{datasets}
