% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicDistrict.R
\docType{data}
\name{BasicDistrict}
\alias{BasicDistrict}
\title{Data Set of District information of China}
\format{
A data frame:
\describe{
\item{id}{ id  }
\item{name}{ name in chn }
\item{parent_id}{ parent id }
\item{initial}{ first string in pinyin }
\item{initials}{ short string in pinyin }
\item{pinyin}{ full string in pinyin }
\item{extra}{ extra detail  }
\item{suffix}{ suffix of district }
\item{code}{ code of district }
\item{area_code}{ area code of district }
\item{order}{ order index }
}
}
\source{
\url{https://www.huhuaping.com/}
}
\usage{
BasicDistrict
}
\description{
A data set containing district information of china from the github
repository 'wecatch/china_regions'\url{https://github.com/eduosi/district},
with wide data format.
}
\keyword{datasets}
