% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-province.R
\docType{data}
\name{BasicProvince}
\alias{BasicProvince}
\title{Data Set of Basic Province information}
\format{
A data frame:
\describe{
\item{id}{ identifier with digits 1 }
\item{province}{ names of province in chinese }
\item{region_pro}{ region of pvovince in chinese }
}
}
\source{
\url{https://www.huhuaping.com/}
}
\usage{
BasicProvince
}
\description{
A data set containing basic information of province and its region,
with wide data format.
}
\keyword{datasets}
