% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicTelAreaCode.R
\docType{data}
\name{BasicTelAreaCode}
\alias{BasicTelAreaCode}
\title{Data Set of Telephone area code information of China}
\format{
A data frame:
\describe{
\item{area_code}{ area tel code }
\item{province_raw}{ name of province in chn }
\item{district}{ name of district in city or county }
}
}
\source{
\url{https://www.huhuaping.com/}
}
\usage{
BasicTelAreaCode
}
\description{
A data set containing Telephone area code information of china from the github
repository 'wecatch/china_regions'\url{https://github.com/eduosi/district},
with wide data format.
}
\keyword{datasets}
