% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicTelcode.R
\docType{data}
\name{BasicTelcode}
\alias{BasicTelcode}
\title{Data Set of Telephone area code information of China}
\format{
A data frame:
\describe{
\item{name}{ name of city and county }
\item{code}{ code of city and county }
\item{area_code}{ code of telephone }
\item{code_province}{ code of province }
\item{district}{ name of province }
}
}
\source{
\url{https://www.huhuaping.com/}
}
\usage{
BasicTelcode
}
\description{
A data set containing Telephone area code information of china from the github
repository 'wecatch/china_regions'\url{https://github.com/eduosi/district},
with wide data format.
}
\keyword{datasets}
