% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_dirs_vec.R
\name{gen_dirs_vec}
\alias{gen_dirs_vec}
\title{Generate batch directories}
\usage{
gen_dirs_vec(media, final)
}
\arguments{
\item{media}{character. First part of director path, with dash '/' at the end}

\item{final}{vector. Bundle of  final directory name, without dash '/'.}
}
\description{
Generate batch directories
}
\examples{
\dontrun{
dir_final <- c("01-machine",
  "02-fertilizer",
  "03-plastic",
  "04-pesticide",
  "05-test")
  dir_media <- "data-raw/rural-yearbook/part03-agri-produce/"

gen_dirs_vec(media = dir_media, final = dir_final)
}

}
