% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_funs.R
\name{get.best.match}
\alias{get.best.match}
\title{Helper function to get the best matched of specified character vectors.}
\usage{
get.best.match(word, charvec)
}
\arguments{
\item{word}{character. The word to match.}

\item{charvec}{character vector. The vector of words to match against.}
}
\value{
character. The best matched word from charvec.
}
\description{
Helper function to get the best matched of specified character vectors.
}
\examples{
\dontrun{
get.best.match("农业机械", c("农业机械", "农业机械总动力"))
}

}
\keyword{internal}
