% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-get-province-of-institution.R
\name{get_province_of_institution}
\alias{get_province_of_institution}
\title{Helper Function to match and get province of target institution}
\usage{
get_province_of_institution(df, target_institution, target_province)
}
\arguments{
\item{df, }{data frame, contains the target institution column}

\item{target_institution, }{character, the column name of target institution}

\item{target_province, }{character, the target output column name for province variable.}
}
\value{
out
}
\description{
Helper Function to match and get province of target institution
}
\examples{
library(magrittr)
utils::data("PubObsStation",package = "techme")
tbl_read <- PubObsStation \%>\%
  dplyr::select(-province)

tbl_out <- get_province_of_institution(df = tbl_read,
  target_institution ="institution",
  target_province = "province")

head(tbl_out)
}
