% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vars.R
\name{get_vars}
\alias{get_vars}
\title{Get names from basic variables table}
\usage{
get_vars(df, lang = "eng", block, what = "variables")
}
\arguments{
\item{df}{data frame}

\item{lang}{characer. Which is the language of variables selected, default "eng",
other options is "chn".}

\item{block}{list. The list must be length 4, with its names following: "block1","block2",
"block3", "block4".}

\item{what}{character. The options are including: "variables"(default),"short_chn",
"short_eng"}
}
\value{
list
}
\description{
Get names from basic variables table
}
\examples{
options(encoding = "UTF-8")
block_sel <- list(block1 ="v7",
block2="sctj",
block3 ="nyjx",
block4= c("zdl","dztlj","xtlj",
"dztlj_pt","xtlj_pt",
"pgddj","pgcyj"))

vars_set <-  get_vars(df = varsList, block = block_sel, what = "variables")

}
