% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-gg2xlsx.R
\name{gg2xlsx}
\alias{gg2xlsx}
\title{Export chunk plot to editable xlsx file}
\usage{
gg2xlsx(p, dir = "pic/ggsave-xlsx/", num_section = 2, num_fig, names_chunk)
}
\arguments{
\item{p}{ggplot object.}

\item{dir}{character. relative dir path string}

\item{num_section}{integer. default 2}

\item{num_fig}{character.}

\item{names_chunk}{character. name  of chunk, from which the plot generateed.}
}
\value{
output xlsx file
}
\description{
Export chunk plot to editable xlsx file
}
\examples{
\dontrun{
  p0 <- plot(1:10, 1:10)
  gg2xlsx(p = p0, dir="pic/ggsave-xlsx/",
        num_section = 5,num_fig = 2,
        names_chunk = "ggplot-chunk")
}
}
