% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_funs.R
\name{help.document}
\alias{help.document}
\title{Help Document the Variables List of Data Set for Package Development}
\usage{
help.document(df)
}
\arguments{
\item{df}{data.frame. The data frame to document.}
}
\value{
NULL. The function prints the documentation template to the console.
}
\description{
This function generates documentation template for data frame variables in Roxygen2 format.
It creates a list of variable entries that can be used in the \code{@format} section
of data documentation.
}
\details{
The function takes a data frame as input and generates documentation entries for each variable
in the format required by Roxygen2. This is particularly useful when documenting datasets
in R packages.
}
\examples{
\dontrun{
data(ProvinceCity)
help.document(ProvinceCity)
}

}
\keyword{internal}
