% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tech-IndustryOperation.R
\docType{data}
\name{IndustryOperation}
\alias{IndustryOperation}
\title{Data Set of Hi-tech Industry Operation}
\format{
A data frame:
\describe{
\item{province}{province, including nation}
\item{year}{year, from 2016 to report year, and no report in year 2017}
\item{chn_block4}{vars name, in chinese}
\item{value}{value of variables}
\item{units}{units of variables}
\item{variables}{vars name, by coding}
}
}
\source{
\url{https://www.huhuaping.com/}
}
\usage{
IndustryOperation
}
\description{
A data set containing Hi-tech Industry Operation from part IV of the Yearbook of Tech,
with long data format.
}
\keyword{datasets}
