% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tech-IndustryTrade.R
\docType{data}
\name{IndustryTrade}
\alias{IndustryTrade}
\title{Date Set of Hi-tech Industry Trade}
\format{
A data frame:
\describe{
\item{province}{province, including nation}
\item{year}{year, from 2016 to 2018, and no report after 2018}
\item{chn_block4}{vars name, in chinese}
\item{value}{value of variables}
\item{units}{units of variables}
\item{variables}{vars name, by coding}
}
}
\source{
\url{https://www.huhuaping.com/}
}
\usage{
IndustryTrade
}
\description{
A data set containing Hi-tech Industry Trade from part IV of the Yearbook of Tech,
with long data format.
}
\keyword{datasets}
