% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Livestock-Breeding.R
\docType{data}
\name{LivestockBreeding}
\alias{LivestockBreeding}
\title{Date Set of Livestock Breeding}
\format{
A data frame:
\describe{
\item{province}{province, including nation}
\item{year}{year, begin with year 2010}
\item{chn_block4}{vars name, in chinese}
\item{value}{value of variables}
\item{units}{units of variables}
\item{variables}{vars name, by coding}
}
}
\source{
\url{https://www.huhuaping.com/}
}
\usage{
LivestockBreeding
}
\description{
A data set containing Livestock Breeding statistics from the Yearbook of Livestock,
with long data format.
}
\keyword{datasets}
