% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tech-MarketPull.R
\docType{data}
\name{MarketPull}
\alias{MarketPull}
\title{Date Set of Tech Market Trade (Pull Effect)}
\format{
A data frame:
\describe{
\item{province}{province, including nation}
\item{year}{year, begin with year 2010}
\item{chn_block4}{vars name, in chinese}
\item{value}{value of variables}
\item{units}{units of variables}
\item{variables}{vars name, by coding}
}
}
\source{
\url{https://www.huhuaping.com/}
}
\usage{
MarketPull
}
\description{
A data set containing Tech Market Trade from the Yearbook of Tech,
with long data format.
}
\keyword{datasets}
