% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-ProvinceCity.R
\docType{data}
\name{ProvinceCity}
\alias{ProvinceCity}
\title{Data Set of Province and City of China}
\format{
A data frame:
\describe{
\item{index}{ ordered index }
\item{province}{ name of province  }
\item{city}{ city belongs of the province }
\item{id}{ id of city with digits 12 }
\item{province_clean}{ reduced province name in chinese }
\item{city_clean}{ reduced city name in chinese  }
}
}
\source{
\url{https://www.huhuaping.com/}
}
\usage{
ProvinceCity
}
\description{
A data set containing Province and City of china from the github
repository 'wecatch/china_regions'\url{https://github.com/wecatch/china_regions},
with wide data format.
}
\keyword{datasets}
