% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pub-AgriParkEval.R
\docType{data}
\name{PubAgriParkEval}
\alias{PubAgriParkEval}
\title{Details of Evaluation result of National Agricultural Sci-tech Park}
\format{
A data frame:
\describe{
\item{year}{ integer, year of document }
\item{index}{ integer, the ordered index of list }
\item{name}{ character, name of the agri park }
\item{result}{ character, evaluation result marked in c('good','ok','fail') }
\item{province}{character, province of agri park in reduced chinese}
}
}
\source{
\url{https://www.huhuaping.com/}
}
\usage{
PubAgriParkEval
}
\description{
A data set containing Evaluation result of National Agricultural Sci-tech Park
year by year from the public
site \url{https://www.most.gov.cn},
with wide data format.
}
\keyword{datasets}
