% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pub-AgriParkList.R
\docType{data}
\name{PubAgriParkList}
\alias{PubAgriParkList}
\title{Details of Approved List of National Agricultural Sci-tech Park}
\format{
A data frame:
\describe{
\item{index}{ integer, the ordered index of list }
\item{batch}{ character, approved batch Number with digits 2 }
\item{name}{ character, name of the agri park }
\item{province}{character, province of agri park in reduced chinese}
}
}
\source{
\url{https://www.huhuaping.com/}
}
\usage{
PubAgriParkList
}
\description{
A data set containing Approved List of National Agricultural Sci-tech Park
year by year from the public
site \url{https://www.most.gov.cn},
with wide data format.
}
\keyword{datasets}
