% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubConvergenceCluster.R
\docType{data}
\name{PubConvergenceCluster}
\alias{PubConvergenceCluster}
\title{Details of Approved List of Special and Advantage Industry Cluster}
\format{
A data frame:
\describe{
\item{index}{ integer, the ordered index of list }
\item{year}{ integer, the official approved year }
\item{name}{ character, name of the Industry Cluster }
\item{province}{character, province of Industry Cluster in reduced chinese}
}
}
\source{
\url{https://www.huhuaping.com/}
}
\usage{
PubConvergenceCluster
}
\description{
A data set containing Approved List of Special and Advantage Industry Cluster
year by year from the "moa-industry-convergence" project
which public at \url{https://www.moa.gov.cn},
with wide data format.
}
\keyword{datasets}
