% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubConvergencePark.R
\docType{data}
\name{PubConvergencePark}
\alias{PubConvergencePark}
\title{Details of Approved List of National Agricultural Demonstration Park}
\format{
A data frame:
\describe{
\item{index}{ integer, the ordered index of list }
\item{year}{ integer, the official approved year }
\item{name}{ character, name of the agri park }
\item{province}{character, province of agri park in reduced chinese}
}
}
\source{
\url{https://www.huhuaping.com/}
}
\usage{
PubConvergencePark
}
\description{
A data set containing Approved List of National Agricultural Demonstration Park
year by year from the "moa-industry-convergence" project
which public at \url{https://www.moa.gov.cn},
with wide data format.
}
\keyword{datasets}
