% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubFreshKeepCounty.R
\docType{data}
\name{PubFreshKeepCounty}
\alias{PubFreshKeepCounty}
\title{Details of Approved List of Spot Agri-products Fresh keeping test County}
\format{
A data frame:
\describe{
\item{index}{ integer, the ordered index of list }
\item{year}{ integer, the official approved year }
\item{name}{ character, name of the approved county }
\item{province}{character, province of approved county in reduced chinese}
}
}
\source{
\url{https://www.huhuaping.com/}
}
\usage{
PubFreshKeepCounty
}
\description{
A data set containing Approved List of Spot Agri-products Fresh keeping test County
year by year from the "moa-freshkeep-county" project
which public at \url{https://www.moa.gov.cn},
with wide data format.
}
\keyword{datasets}
