% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tech-PublicBudget.R
\docType{data}
\name{PublicBudget}
\alias{PublicBudget}
\title{Data Set of Public Budget}
\format{
A data frame:
\describe{
\item{province}{province, including nation}
\item{year}{year, begin with year 2010}
\item{chn_block4}{vars name, in chinese}
\item{value}{value of variables}
\item{units}{units of variables}
\item{variables}{vars name, by coding}
}
}
\source{
\url{https://www.huhuaping.com/}
}
\usage{
PublicBudget
}
\description{
A data set containing Public Budget statistics from the Yearbook of Nation,
with long data format.
}
\keyword{datasets}
