% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pub-ObsStation.R
\docType{data}
\name{PubObsStation}
\alias{PubObsStation}
\title{Details list of Observe Stations of MOA and MOE}
\format{
A data frame:
\describe{
\item{officer}{ name of Ministry, MOA or MOE }
\item{year}{ integer, year of the document from Ministry site }
\item{index}{ integer, the ordered index of list }
\item{name}{ character, name of the agri park }
\item{institution}{ character, name  of institution hosted the stations }
\item{province}{character, province of agri park in reduced chinese}
}
}
\source{
\url{https://www.huhuaping.com/}
}
\usage{
PubObsStation
}
\description{
A data set containing list of Observe Stations of Ministry of Agriculture(MOA)
and Ministry of Education (MOE)
year by year from the public
MOA site \url{http://www.moa.gov.cn/} and
MOE site \url{http://www.moe.gov.cn/} ,
with wide data format.
}
\keyword{datasets}
