% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubRuralInfoBase.R
\docType{data}
\name{PubRuralInfoBase}
\alias{PubRuralInfoBase}
\title{Details of Approved List of Agricultural and Rural informational Demonstration Base}
\format{
A data frame:
\describe{
\item{year}{ integer, the official approved year }
\item{index}{ integer, the ordered index of list }
\item{type}{character, type of the approved base }
\item{institution}{ character, name of the approved base }
\item{province}{character, province of approved base in reduced chinese}
}
}
\source{
\url{https://www.huhuaping.com/}
}
\usage{
PubRuralInfoBase
}
\description{
A data set containing Approved List of Agricultural and Rural informational Demonstration Base
year by year from the "moa-ruralinfo-base" project
which public at \url{https://www.moa.gov.cn},
with wide data format.
}
\keyword{datasets}
