% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pub-SKLMost.R
\docType{data}
\name{PubSKLMost}
\alias{PubSKLMost}
\title{Details of Stated Key Lab (SKL)}
\format{
A data frame:
\describe{
\item{year}{ year of data source  }
\item{type}{ type of SKL, c('SKL','PSKL', 'FSKL')  }
\item{index}{ index }
\item{name}{ name of SKL }
\item{area}{ research areas of SKL }
\item{institution}{ institution of SKL }
\item{administrator}{ administrator of SKL }
\item{province}{ province of SKL }
\item{NO}{ NO of SKL  }
\item{start_year}{ start year of SKL  }
}
}
\source{
\url{https://www.huhuaping.com/}
}
\usage{
PubSKLMost
}
\description{
A data set containing Approved List of Stated Key Lab
(SKL) year by year from the public
site \url{https://www.most.gov.cn},
with wide data format.
}
\keyword{datasets}
