% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tech-RDIntense.R
\docType{data}
\name{RDIntense}
\alias{RDIntense}
\title{Date Set of RD Intense}
\format{
A data frame:
\describe{
\item{province}{province, including nation}
\item{year}{year, begin with year 2010}
\item{chn_block4}{vars name, in chinese}
\item{value}{value of variables}
\item{units}{units of variables}
\item{variables}{vars name, by coding}
}
}
\source{
\url{https://www.huhuaping.com/}
}
\usage{
RDIntense
}
\description{
A data set containing RD Intense statistics from the Yearbook of Tech,
with long data format.
}
\keyword{datasets}
