% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/techme-package.R
\docType{package}
\name{techme-package}
\alias{techme}
\alias{techme-package}
\title{techme: Generate Research Report with Complex Data Sources}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/huhuaping/techme}
  \item \url{https://huhuaping.github.io/techme/}
  \item Report bugs at \url{https://github.com/huhuaping/techme/issues}
}

}
\author{
\strong{Maintainer}: Kevin Hu \email{huhuaping01@hotmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Kevin Hu \email{huhuaping01@hotmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
}

}
\keyword{internal}
