% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_funs.R
\name{wfl.tidyTable}
\alias{wfl.tidyTable}
\title{Tidy dataset from unpivot yearbook table}
\usage{
wfl.tidyTable(dt)
}
\arguments{
\item{dt}{data.frame. typically the yearbook unpivot table by use \code{unpivot}.}
}
\value{
data.frame
}
\description{
Tidy dataset from unpivot yearbook table
}
\examples{
\dontrun{
df_tidy <- wfl.tidyTable(dt = df_out)
}

}
\keyword{internal}
