% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_funs.R
\name{wfl.useData}
\alias{wfl.useData}
\title{Read the raw xlsx files and use the specified data.frame by use_data()}
\usage{
wfl.useData(directory.source, file.pattern, name.dt, which.dt = "df_use")
}
\arguments{
\item{directory.source}{character. The path of the source xlsx files.}

\item{file.pattern}{character. The pattern of the file name.}

\item{name.dt}{character. The name of the data frame.
And the name has uniform format style as "AgriMachine", "LivestockBreeding", etc.}

\item{which.dt}{character. The name of the data frame to use, default is "df_use".}
}
\description{
Read the raw xlsx files and use the specified data.frame by use_data()
}
\examples{
\dontrun{
wfl.useData(
    directory.source = "data-raw/data-tidy/",
    file.pattern = "\\\\d{4}",
    name.dt = "AgriMachine",
    which.dt = "df_use"
)
}

}
\keyword{internal}
