% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_funs.R
\name{wfl.writeXlsx}
\alias{wfl.writeXlsx}
\title{Write out xlsx with the output file which the file name is specified year and prefix label}
\usage{
wfl.writeXlsx(dt, file_source = file_xlsx, year_target, prefix_label = NULL)
}
\arguments{
\item{dt}{data.frame. The data frame to write out.}

\item{file_source}{character. The path of the source file.}

\item{year_target}{numeric. The year to write out.}

\item{prefix_label}{character. The prefix label of the file name, default is NULL. Other value may be "funds", "ammount", etc.}
}
\description{
Write out xlsx with the output file which the file name is specified year and prefix label
}
\examples{
\dontrun{
wfl.writeXlsx(
    dt = df_matched,
    file_source = "data-raw/df-matched.xlsx",
    year_target = 2019,
    prefix_label = "funds"
)
}

}
\keyword{internal}
