% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_funs.R
\name{wfl.Xls2Xlsx}
\alias{wfl.Xls2Xlsx}
\title{Convert protected xls file to xlsx file, and remove unnecessary sheet.
This function will remove the xls file permission protected by CNKI.}
\usage{
wfl.Xls2Xlsx(file_path, sheet_drop = c("CNKI"))
}
\arguments{
\item{file_path}{character. Path to the target xls file}

\item{sheet_drop}{character. Name of the sheet to drop}
}
\value{
character. Path to the converted xlsx file
}
\description{
Convert protected xls file to xlsx file, and remove unnecessary sheet.
This function will remove the xls file permission protected by CNKI.
}
\examples{
\dontrun{
wfl.Xls2Xlsx(
    file_path = "data-raw/example.xls",
    sheet_drop = c("CNKI")
)
}

}
\keyword{internal}
