% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_funs.R
\name{workflow_funs}
\alias{workflow_funs}
\alias{getRange}
\alias{unpivot}
\alias{getInfo}
\alias{wfl.unpivotXlsx}
\alias{wfl.tidyTable}
\alias{get.targetList}
\alias{get.vars}
\alias{get.best.match}
\alias{wfl.matchVars}
\alias{wfl.addVars}
\alias{wfl.writeXlsx}
\alias{choose.nameData}
\alias{wfl.useData}
\alias{help.document}
\title{Workflow Functions:}
\usage{
getRange(dt, ith, what, reg_start, reg_end)

unpivot(dt, rows, cols, cols.drop = NULL, header.mode, vars.add = NULL)

getInfo(dt, unit_pattern)

wfl.unpivotXlsx(
  file,
  header.mode = "vars-year",
  vars.add = NULL,
  cols.drop = NULL,
  pattern.table,
  reg_start,
  reg_end,
  unit_pattern
)

wfl.tidyTable(dt)

get.targetList()

get.vars(df, lang = "eng", block, what = "variables")

get.best.match(word, charvec)

wfl.matchVars(dt, block_target, block_lang = "eng")

wfl.addVars(dt_left, dt_right)

wfl.writeXlsx(dt, file_source = file_xlsx, year_target, prefix_label = NULL)

choose.nameData()

wfl.useData(directory.source, file.pattern, name.dt, which.dt = "df_use")

help.document(df)
}
\arguments{
\item{dt}{data.frame. The data frame to write out.}

\item{ith}{number. when exist multiple table region in one sheet.}

\item{what}{character. The type of variable names to return. Options include:
\itemize{
\item "variables" (default): Full variable names
\item "short_chn": Short Chinese names
\item "short_eng": Short English names
\item "chn_block4": Chinese block4 names
\item "eng_block4": English block4 names
\item other: other variable names in the input data frame (typically \code{varList})
}}

\item{reg_start}{character. regex pattern for start identifier.}

\item{reg_end}{character. regex pattern for end identifier.}

\item{rows}{vector. Target rows of the region contains pivot table.}

\item{cols}{vector. Target cols of the region contains pivot table.}

\item{cols.drop}{vector. Columns numbers which will be dropped, default \code{NULL}.}

\item{header.mode}{character. One of the four options:  'vars-year', 'vars', 'vars-vars','year'}

\item{vars.add}{character. if header.mode = 'year', then the vars.add must be specified,
And you can use the function \code{get_vars()} to get the variable name.}

\item{file}{character. Path to the xlsx file.}

\item{pattern.table}{character. Regex pattern for table identifier.}

\item{df}{data.frame. The data frame to document.}

\item{lang}{character. The language of variables to select. Must be either "eng" (default)
or "chn".}

\item{block}{list. A list which the length is not greater than 4 with names "block1", "block2", "block3", "block4".
Each element contains the block identifiers to filter by. Note that block4 is optional.}

\item{word}{character. The word to match.}

\item{charvec}{character vector. The vector of words to match against.}

\item{block_target}{list. Names of the list should be part or all of: block1, block2, block3, block4.}

\item{block_lang}{character. Specify which language to use, either 'eng'(default) or 'chn'.}

\item{dt_left}{data.frame. The tidy unpivot table.}

\item{dt_right}{data.frame. The matched variables table and should be checked.}

\item{file_source}{character. The path of the source file.}

\item{year_target}{numeric. The year to write out.}

\item{prefix_label}{character. The prefix label of the file name, default is NULL. Other value may be "funds", "ammount", etc.}

\item{directory.source}{character. The path of the source xlsx files.}

\item{file.pattern}{character. The pattern of the file name.}

\item{name.dt}{character. The name of the data frame.
And the name has uniform format style as "AgriMachine", "LivestockBreeding", etc.}

\item{which.dt}{character. The name of the data frame to use, default is "df_use".}
}
\value{
vector

data.frame

vector. A character vector containing the unit information.

data.frame. A data frame containing the unpivoted data with units information.

data.frame

list. List of target list with block identifiers.

data.frame. A data frame containing the filtered variable names.

character. The best matched word from charvec.

data.frame. A data frame containing the matched variables.

data.frame. A data frame containing the joined and processed data.

character or NULL. The selected data name from the list, or NULL if option 0 is selected.

NULL. The function prints the documentation template to the console.
}
\description{
Internal functions for handling various workflow tasks in the techme package.

This function provides an interactive way to select a target list from a predefined collection
used for filtering and organizing data in the package. The lists are organized by different
categories (v4, v6, v7, v8) and contain block identifiers for data filtering.

This function retrieves variable names from a basic variables table based on specified
language and block criteria. It supports filtering variables by hierarchical block
structure in both English and Chinese.

This function provides an interactive way to select a data name from a predefined list.
It displays all available options and allows the user to choose by entering a number.

This function generates documentation template for data frame variables in Roxygen2 format.
It creates a list of variable entries that can be used in the \code{@format} section
of data documentation.
}
\details{
This file contains a collection of internal functions for:
\itemize{
\item function \code{getRange()}: get the range of the pivot table in the xlsx file
\item function \code{unpivot()}: unpivot the xlsx file
\item function \code{getInfo()}: get the information of the xlsx file
\item function \code{wfl.unpivotXlsx()}: unpivot the xlsx file
}

This function is used to get the range of the pivot table in the xlsx file.
The pivot table is identified by the regex pattern for start and end identifier.
The function will return the row and column range of the target pivot table.
Note 1: there may be only one table or multiple tables in the sheet
Note 2: when there are multiple tables, these table's alignment may be horizontal or vertical
Note 3: We only assume the only multiple table's alignment is horizontal or vertical, not hybrid alignment.

This function is used to loop unpivot all the sheets in the xlsx file.
It will call three internal functions respectively:
\enumerate{
\item getRange(): to get the range of the pivot table in the xlsx file.
\item unpivot(): to unpivot the pivot table in the xlsx file.
\item getInfo(): to get the unit information of the pivot table in the xlsx file.
Keep in mind that a sheet may contains multiple pivot tables.
Note 1: there may be only one table or multiple tables in the sheet
Note 2: when there are multiple tables, these table's alignment may be horizontal or vertical
Note 3: We only assume the only multiple table's alignment is horizontal or vertical, not hybrid alignment.
}

The target lists are organized into several categories:
\itemize{
\item v4: Research and Development related lists (RDnbs, RDinner, RDfirm, etc.)
\item v6: Budget related lists
\item v7: Agricultural production related lists (machine, fertilizer, plastic, pesticide)
\item v8: Livestock related lists (t1-t9)
}
Each list contains block identifiers (block1, block2, block3) used for filtering data.

The function filters variables based on a hierarchical block structure (block1-4)
and returns the specified variable names. It supports both English and Chinese
variable names through the language parameter. The block4 parameter is optional
in the block list.

The function takes a data frame as input and generates documentation entries for each variable
in the format required by Roxygen2. This is particularly useful when documenting datasets
in R packages.
}
\examples{
\dontrun{
wb <- loadWorkbook(path_xls, create = TRUE)
dt <- readWorksheet(wb, sheet = 1, header = F) \%>\%
    select_if(~ !all(is.na(.)))
i <- 2
myrows <- getRange(dt, ith = i, what = "row")
mycols <- getRange(dt, ith = i, what = "col")
}

\dontrun{
wb <- loadWorkbook(path_xls, create = TRUE)
dt <- readWorksheet(wb, sheet = 1, header = F) \%>\%
    select_if(~ !all(is.na(.)))
i <- 1
myrows <- getRange(dt, ith = i, what = "row")
mycols <- getRange(dt, ith = i, what = "col")
cols_drop <- c(2)
header_mode <- c("vars", "vars-vars")
vars_spc <- get_vars(
    df = varsList, lang = "eng",
    block = list(
        block1 = "v4",
        block2 = "zh",
        block3 = "qd",
        block4 = "RD"
    ),
    what = "chn_block4"
)

tbl_raw <- unpivot(dt,
    cols = mycols, rows = myrows,
    cols.drop = cols_drop,
    header.mode = header_mode[i],
    vars.add = vars_spc
)
}

\dontrun{
same_units <- getInfo(dt)
}

\dontrun{
wfl.unpivot_xlsx(
    file = "data-raw/example.xlsx",
    header.mode = "vars-year",
    cols.drop = NULL,
    pattern.table = "^\\\\u5730.*\\\\u533a"
)
}

\dontrun{
df_tidy <- wfl.tidyTable(dt = df_out)
}

\dontrun{
# Interactive selection of target list
target_list <- get.targetList()
}

\dontrun{
options(encoding = "UTF-8")
block_sel <- list(
    block1 = "v7",
    block2 = "sctj",
    block3 = "nyjx"
)

vars_set <- get.vars(df = varsList, block = block_sel, what = "variables")
}

\dontrun{
get.best.match(
    "\u519c\u4e1a\u673a\u68b0",
    c(
        "\u519c\u4e1a\u673a\u68b0",
        "\u519c\u4e1a\u673a\u68b0\u603b\u52a8\u529b"
    )
)
}

\dontrun{
target <- list(block1 = "v7", block2 = "sctj", block3 = "nyjx")
df_vars_matched <- wfl.matchVars(dt = df_tidy, block_target = target)
}

\dontrun{
df_matched <- wfl.addVars(dt_left = df_tidy, dt_right = vars_matched)
}

\dontrun{
wfl.writeXlsx(
    dt = df_matched,
    file_source = "data-raw/df-matched.xlsx",
    year_target = 2019,
    prefix_label = "funds"
)
}

\dontrun{
# Interactive selection of data name
selected_name <- choose.nameData()
}

\dontrun{
wfl.useData(
    directory.source = "data-raw/data-tidy/",
    file.pattern = "\\\\d{4}",
    name.dt = "AgriMachine",
    which.dt = "df_use"
)
}

\dontrun{
data(ProvinceCity)
help.document(ProvinceCity)
}

}
\keyword{internal}
