% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_funs.R
\name{workflow_funs_use_data}
\alias{workflow_funs_use_data}
\alias{choose.nameData}
\alias{wfl.useData}
\title{Workflow Functions: use data}
\usage{
choose.nameData()

wfl.useData(directory.source, file.pattern, name.dt, which.dt = "df_use")
}
\arguments{
\item{directory.source}{character. The path of the source xlsx files.}

\item{file.pattern}{character. The pattern of the file name.}

\item{name.dt}{character. The name of the data frame.
And the name has uniform format style as "AgriMachine", "LivestockBreeding", etc.}

\item{which.dt}{character. The name of the data frame to use, default is "df_use".}
}
\value{
character or NULL. The selected data name from the list, or NULL if option 0 is selected.
}
\description{
Internal functions for handling various workflow tasks in the techme package.

This function provides an interactive way to select a data name from a predefined list.
It displays all available options and allows the user to choose by entering a number.
}
\details{
This file contains a collection of internal functions for:
\itemize{
\item function \code{choose.nameData()}: choose the name of \code{use_data()}
\item function \code{wfl.useData()}: use the data with \code{use_data()}
}
}
\examples{
\dontrun{
# Interactive selection of data name
selected_name <- choose.nameData()
}

\dontrun{
wfl.useData(
    directory.source = "data-raw/data-tidy/",
    file.pattern = "\\\\d{4}",
    name.dt = "AgriMachine",
    which.dt = "df_use"
)
}

}
\keyword{internal}
