% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-add-footer.R
\name{add_footer_asis}
\alias{add_footer_asis}
\title{Add success footer information as whole part for the figure or table}
\usage{
add_footer_asis(note, source)
}
\arguments{
\item{note}{character, note text for figure or table.}

\item{source}{character, source text for figure or table.}
}
\value{
invisible
}
\description{
This function should be in the dependent code chunk (outside the code chunk of plot or kable)
and must set the chunk options "result= 'asis'".
}
\examples{

cap_note <- "This is a note, ba la ba la"
cap_source <- "source from yearbook 2021"
\dontrun{
 add_footer_asis(cap_note, cap_source)
}

}
