% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-docDataset.R
\name{document_dt}
\alias{document_dt}
\title{Help Document the Variables List of Data Set for Package Development}
\usage{
document_dt(df)
}
\arguments{
\item{df}{a data frame}
}
\description{
Help Document the Variables List of Data Set for Package Development
}
\examples{
data(ProvinceCity)
document_dt(ProvinceCity)


}
