% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pub-StandardXmj.R
\docType{data}
\name{PubStandardXmj}
\alias{PubStandardXmj}
\title{Details of Officer Livestock Standard List from MOA (Xmj)}
\format{
A data frame:
\describe{
\item{year}{integer, year marked in the file name}
\item{index}{character,  ordered index number }
\item{province}{ character, province of origin institution  }
\item{prod_name}{character, livestock type }
\item{com_name}{ name of the company }
}
}
\source{
\url{https://www.huhuaping.com/}
}
\usage{
PubStandardXmj
}
\description{
A data set containing detail information of Officer Livestock Standard List
from the public site MOA (Xmj) \url{http://www.moa.gov.cn},
with wide data format.
}
\keyword{datasets}
