% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_funs.R
\name{workflow_funs}
\alias{workflow_funs}
\alias{wfl.findFiles}
\alias{wfl.Xls2Xlsx}
\alias{getRange}
\alias{unpivot}
\alias{getInfo}
\alias{wfl.unpivotXlsx}
\alias{wfl.tidyTable}
\alias{wfl.matchVars}
\alias{wfl.addVars}
\alias{wfl.writeXlsx}
\title{Workflow Functions}
\usage{
wfl.findFiles(directory, pattern)

wfl.Xls2Xlsx(file_path, sheet_drop = c("CNKI"))

getRange(dt, ith, what, reg_start, reg_end)

unpivot(dt, rows, cols, cols.drop = NULL, header.mode, vars.add = NULL)

getInfo(dt, unit_pattern)

wfl.unpivotXlsx(
  file,
  header.mode = "vars-year",
  vars.add = NULL,
  cols.drop = NULL,
  pattern.table,
  reg_start,
  reg_end,
  unit_pattern
)

wfl.tidyTable(dt)

wfl.matchVars(dt, block_target, block_lang = "eng")

wfl.addVars(dt_left, dt_right)

wfl.writeXlsx(dt, file_source = file_xlsx, year_target, prefix_label = NULL)
}
\arguments{
\item{directory}{character. Path to search for files}

\item{pattern}{character. Pattern to match files}

\item{file_path}{character. Path to the target xls file}

\item{sheet_drop}{character. Name of the sheet to drop}

\item{dt}{data.frame. The data frame to write out.}

\item{ith}{number. when exist multiple table region in one sheet.}

\item{what}{character. What is the object function will return, whether "row" or "col".}

\item{reg_start}{character. regex pattern for start identifier, default \code{'^地.*区'}.}

\item{reg_end}{character. regex pattern for end identifier, default \code{'^新.*疆'}.}

\item{rows}{vector. Target rows of the region contains pivot table.}

\item{cols}{vector. Target cols of the region contains pivot table.}

\item{cols.drop}{vector. Columns numbers which will be dropped, default \code{NULL}.}

\item{header.mode}{character. One of the four options:  'vars-year', 'vars', 'vars-vars','year'}

\item{vars.add}{character. if header.mode = 'year', then the vars.add must be specified,
And you can use the function \code{get_vars()} to get the variable name.}

\item{file}{character. Path to the xlsx file.}

\item{pattern.table}{character. Regex pattern for table identifier, default \code{'续表'}.}

\item{block_target}{list. Names of the list should be part or all of: block1, block2, block3, block4.}

\item{block_lang}{character. Specify which language to use, either 'eng'(default) or 'chn'.}

\item{dt_left}{data.frame. The tidy unpivot table.}

\item{dt_right}{data.frame. The matched variables table and should be checked.}

\item{file_source}{character. The path of the source file.}

\item{year_target}{numeric. The year to write out.}

\item{prefix_label}{character. The prefix label of the file name, default is NULL. Other value may be "funds", "ammount", etc.}
}
\value{
character vector of file paths

character. Path to the converted xlsx file

vector

data.frame

vector. A character vector containing the unit information.

data.frame. A data frame containing the unpivoted data with units information.

data.frame

data.frame. A data frame containing the matched variables.

data.frame. A data frame containing the joined and processed data.
}
\description{
Internal functions for handling various workflow tasks in the techme package.
}
\details{
This file contains a collection of internal functions for:
\itemize{
\item Converting and processing Excel files
\item Unpivoting data tables
\item Matching and standardizing variables
\item Writing output files
}
}
\examples{
\dontrun{
wfl.findFiles(directory = "data-raw", pattern = "\\\\.xlsx$")
}

\dontrun{
wfl.Xls2Xlsx(
    file_path = "data-raw/example.xls",
    sheet_drop = c("CNKI")
)
}

\dontrun{
wb <- loadWorkbook(path_xls, create = TRUE)
dt <- readWorksheet(wb, sheet = 1, header = F) \%>\%
    select_if(~ !all(is.na(.)))
i <- 2
myrows <- getRange(dt, ith = i, what = "row")
mycols <- getRange(dt, ith = i, what = "col")
}

\dontrun{
wb <- loadWorkbook(path_xls, create = TRUE)
dt <- readWorksheet(wb, sheet = 1, header = F) \%>\%
    select_if(~ !all(is.na(.)))
i <- 1
myrows <- getRange(dt, ith = i, what = "row")
mycols <- getRange(dt, ith = i, what = "col")
cols_drop <- c(2)
header_mode <- c("vars", "vars-vars")
vars_spc <- get_vars(
    df = varsList, lang = "eng",
    block = list(
        block1 = "v4",
        block2 = "zh",
        block3 = "qd",
        block4 = "RD"
    ),
    what = "chn_block4"
)

tbl_raw <- unpivot(dt,
    cols = mycols, rows = myrows,
    cols.drop = cols_drop,
    header.mode = header_mode[i],
    vars.add = vars_spc
)
}

\dontrun{
same_units <- getInfo(dt)
}

\dontrun{
wfl.unpivot_xlsx(
    file = "data-raw/example.xlsx",
    header.mode = "vars-year",
    cols.drop = NULL,
    pattern.table = "^地.*区"
)
}

\dontrun{
df_tidy <- wfl.tidyTable(dt = df_out)
}

\dontrun{
target <- list(block1 = "v7", block2 = "sctj", block3 = "nyjx")
df_vars_matched <- wfl.matchVars(dt = df_tidy, block_target = target)
}

\dontrun{
df_matched <- wfl.addVars(dt_left = df_tidy, dt_right = vars_matched)
}

\dontrun{
wfl.writeXlsx(
    dt = df_matched,
    file_source = "data-raw/df-matched.xlsx",
    year_target = 2019,
    prefix_label = "funds"
)
}

}
\keyword{internal}
